fix/Расширил компонент. Добавил дочерний компонент input-file-container
