+++ src/components/ui/input-file/index.ts
@@ -1,1 +1,2 @@
 export * from './input-file';
+export * from './input-file-container/input-file-container';
